requirements.txt
loguru==0.7.0
opencv-python==4.7.0.72
optuna==3.1.1
Pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
tensorboard==2.11.0
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tqdm==4.65.0
numpy==1.23.5
prettytable==3.5.0
